pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gym_donkeycar"
dynamic = ["version"]
description = "OpenAI Gym Environments for Donkey Car"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Tawn Kramer", email = "tawnkramer@gmail.com" },
]
keywords = ["donkeycar", "environment", "agent", "rl", "openaigym", "openai-gym", "gym"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "gym==0.22",
    "numpy",
    "pillow",
    "setuptools",  # provides distutils for gym on Python 3.12+
]

[project.optional-dependencies]
tests = [
    "pytest",
    "pytest-mock",
    "pytest-cov",
    "pytest-env",
    "pytest-xdist",
    "pytype",
    "ruff",
    "isort>=5.0",
    "black",
]
docs = [
    "sphinx",
    "sphinx-autobuild",
    "sphinx-rtd-theme",
    "sphinxcontrib-spelling",
    "sphinx-autodoc-typehints",
]
vlm = [
    "torch>=2.0",
    "transformers>=4.45.0",
    "accelerate>=0.25.0",
    "peft>=0.7.0",
    "bitsandbytes>=0.41.0",
    "trl>=0.10.0",
    "datasets>=2.16.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "qwen-vl-utils>=0.0.8",
    "pillow",
]

[project.urls]
Homepage = "https://github.com/Lev-Stambler/gym-donkeycar-uv"
Documentation = "https://gym-donkeycar.readthedocs.io/en/latest/"
Repository = "https://github.com/Lev-Stambler/gym-donkeycar-uv"

[tool.hatch.version]
path = "gym_donkeycar/version.txt"
pattern = "(?P<version>.+)"

[tool.hatch.build.targets.wheel]
packages = ["gym_donkeycar"]

[tool.black]
line-length = 127

[tool.ruff]
line-length = 127

[tool.ruff.mccabe]
max-complexity = 15

[tool.pytype]
inputs = ["gym_donkeycar"]

[tool.isort]
profile = "black"
src_paths = ["gym_donkeycar"]
line_length = 127

[tool.pytest.ini_options]
testpaths = ["tests"]
